Add support for local modifications in ~/.gdbinit.local.
After setting the default options, ~/.gdbinit.local is read, enabling
the user to override the default settings or adding own config parts
without editing the main gdbinit. This should make updates easier.

To allow disabling prompt coloring in that way, the prompt-setting code
was moved as well.

No error or warning is raised if the file does not exist (gdb 7.4.1-debian).