Fixes issues in data printers
When passing an expression containing spaces as an argument to a GDB
command it will be interpreted as multiple arguments even when enclosed
in parentheses. In order to follow the general coding style these
arguments are broken out and set in local variables that are then
passed as arguments instead.