Rename GDB command files to .gdb
Renames the GDB command files to separate them from the shell scripts
etc. Also, this makes it easier for the editor to figure out what
indentation, syntaxt highlighting etc to use.

Also renames the CPU-specific modules to cpu-<architecture>.gdb to make
it more obvious that they are submodules to cpu.gdb.